move collection_resources from intelnalstorage to util
